+++ src/BootstrapProvider/index.js
@@ -1,15 +1,20 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { ThemeProvider } from 'styled-components';
-import themeBs, { makeTheme } from '../theme';
-import UtilityProvider from './UtilityProvider';
+import { ThemeProvider, injectGlobal } from 'styled-components';
+import { getGlobalStyleNoBootstrapProvider } from 'bootstrap-styled-mixins/lib/utilities/reboot';
+import themeBs, { makeTheme as makeThemeBs } from '../theme';
+import UtilityProvider, { defaultProps as utilityProviderDefaultProps } from './UtilityProvider';
 
 export const defaultProps = {
   theme: themeBs,
+  injectGlobal: true,
+  utils: utilityProviderDefaultProps.utils,
 };
 export const propTypes = {
   /** Specified node element will be passed as children of `<BootstrapProvider />` component. */
   children: PropTypes.node,
+  /** Define if BootstrapProvider is in charge of calling getGlobalStyles for you, it can only be run once. */
+  injectGlobal: PropTypes.bool,
   /** Theme variables. */
   theme: PropTypes.object,
   /** Utilities variables. */
@@ -55,16 +60,69 @@ class BootstrapProvider extends React.Component { // eslint-disable-line react/p
   /* eslint-enable react/no-unused-prop-types */
 
   state = {
+    theme: {},
     isWindowPhone8Fixed: null, // just for unit testing
   }
 
-  // componentWillReceiveProps(nextProps) {
-  //   const diff = 0;
-  //
-  // }
+  componentWillMount() {
+    this.setTheme(this.props, this.injectGlobal);
+  }
 
   componentDidMount() {
     this.windowPhone8Fix();
+  }
+
+  /**
+   * On update props
+   * @param nextProps
+   */
+  componentWillReceiveProps(nextProps) {
+    /**
+     * We only replace the theme if the filtered theme is different from the current one
+     */
+    if (JSON.stringify(this.makeTheme(this.props)) !== JSON.stringify(nextProps.theme)) {
+      this.setTheme(this.props);
+    }
+  }
+
+  /**
+   * setTheme is in charge of replacing the theme, it does not filter it, use makeTheme for that
+   * @param userTheme
+   * @param cb
+   */
+  setTheme = ({ theme: userTheme }, cb) => {
+    const theme = this.makeTheme({ theme: userTheme });
+    this.setState({ theme }, cb);
+  }
+
+  /**
+   * makeTheme is in charge of filtering the meta data out of the theme
+   * @param userTheme
+   * @returns {*}
+   */
+  makeTheme = ({ theme: userTheme }) => {
+    const theme = makeThemeBs(userTheme);
+    const metaKeyList = Object.keys(theme).filter((f) => f[0] === '_');
+    metaKeyList.forEach((k) => {
+      delete theme[k];
+    });
+    return theme;
+  }
+
+  injectGlobal = () => {
+    if (this.props.injectGlobal) {
+      /* eslint-disable no-unused-expressions */
+      injectGlobal`
+        ${getGlobalStyleNoBootstrapProvider(
+          this.state.theme['$font-family-base'] || themeBs['$font-family-base'],
+          this.state.theme['$font-size-base'] || themeBs['$font-size-base'],
+          this.state.theme['$font-weight-base'] || themeBs['$font-weight-base'],
+          this.state.theme['$line-height-base'] || themeBs['$line-height-base'],
+          this.state.theme['$body-color'] || themeBs['$body-color'],
+          this.state.theme['$body-bg'] || themeBs['$body-bg'],
+        )}
+      `;
+    }
   }
 
   windowPhone8Fix() {
@@ -84,14 +142,9 @@ class BootstrapProvider extends React.Component { // eslint-disable-line react/p
   }
 
   render() {
-    const { children, theme: userTheme, utils } = this.props;
-    const theme = makeTheme(userTheme);
-    const metaKeyList = Object.keys(theme).filter((f) => f[0] === '_');
-    metaKeyList.forEach((k) => {
-      delete theme[k];
-    });
+    const { children, utils } = this.props;
     return (
-      <ThemeProvider theme={theme}>
+      <ThemeProvider theme={this.state.theme}>
         <UtilityProvider utils={utils}>
           {children}
         </UtilityProvider>