+++ src/Cards/theme.js
@@ -1,7 +1,7 @@
 import Color from 'color';
 import { assertAscending, assertStartAtZero } from 'bootstrap-styled-mixins/lib/variables';
 import unitUtils from 'bootstrap-styled-utils/lib/unitUtils';
-import { allowFalseValue } from '../makeTheme/utils';
+import { allowFalseValue } from '../theme/utils';
 
 const { detectUnit, rmUnit, UNIT } = unitUtils;