+++ src/P/theme.js
@@ -1,5 +1,5 @@
 import unitUtils from 'bootstrap-styled-utils/lib/unitUtils';
-import { allowFalseValue } from '../makeTheme/utils';
+import { allowFalseValue } from '../theme/utils';
 const { rmUnit, UNIT } = unitUtils;
 
 /**