+++ src/Ul/theme.js
@@ -1,5 +1,5 @@
 import { assertAscending, assertStartAtZero } from 'bootstrap-styled-mixins/lib/variables';
-import { allowFalseValue } from '../makeTheme/utils';
+import { allowFalseValue } from '../theme/utils';
 /**
  * BOOTSTRAP THEME UNORDERED LIST
  *