+++ src/makeTheme/index.js
@@ -1,1 +1,0 @@
-export { default, default as makeTheme } from './makeTheme';