+++ src/theme/tests/index.test.js
@@ -1,4 +1,4 @@
-import makeTheme from '../index';
+import { makeTheme } from '../index';
 
 describe('makeTheme exports', () => {
   it('should have makeTheme defined', () => {