+++ CHANGELOG.md
@@ -6,6 +6,22 @@ All notable changes to this project will be documented in this file.
 
 - TODO: write changelog before preparing next tag
 
+## [v1.5.15] - 2018-04-13
+
+- Improve rerendering of the `BootstrapProvider`.
+- Upgrade `react-styleguidedist` to 6.5.0.
+- Rearrange peer / global / dependencies.
+
+## [v1.5.14] - 2018-04-12
+
+- Using `getGlobalStyleNoBootstrapProvider` within `BootstrapProvider` instead of `getGlobalStyles`.
+
+## [v1.5.13] - 2018-04-12
+
+- Rename `makeTheme` to `theme` and removed `theme.js`.
+- Improve `makeTheme`, first arg can accept a list of makeTheme instead of the object override.
+- Using state instead of props in `BootstrapProvider`.
+
 ## [v1.5.12] - 2018-04-12
 
 - Removed argument in `UtilityProvider` based on a mistake in `bootstrap-styled-mixin`.
@@ -25,6 +41,8 @@ All notable changes to this project will be documented in this file.
 
 ## [v1.5.8] - 2018-03-27
 
+=======
+>>>>>>> Stashed changes
 - Change all `getRef` with `innerRef`.
 
 ## [v1.5.7] - 2018-03-27